requirements.txt
azure-cosmos    == 4.7.0
beautifulsoup4  == 4.12.3
boto3           == 1.35.5
python-dotenv   == 1.0.1
requests        == 2.32.3
retry           == 0.9.2
statistics      == 1.0.3.5
xxhash          == 3.5.0
flake8 == 7.1.0
Sphinx == 8.0.2
flask == 3.1.0
sqlmodel
psycopg2